final version of ft_printf

+++ ft_print_string.c
@@ -6,7 +6,7 @@
 /*   By: osallak <osallak@student.1337.ma>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/23 00:08:42 by osallak           #+#    #+#             */
-/*   Updated: 2021/11/24 20:07:32 by osallak          ###   ########.fr       */
+/*   Updated: 2021/11/24 21:08:38 by osallak          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ int	ft_print_string(char *s)
 		write (1, "(null)", 6);
 		return (6);
 	}
-	len = (int) ft_strlen(s);
+	len = ft_strlen(s);
 	while (*s)
 		write(1, s++, 1);
 	return (len);
